requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-multipart==0.0.6

# Document Processing
pymupdf==1.26.6
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Vector DB & Embeddings
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# LLM & RAG
langchain==0.0.350
langchain-community==0.0.2
openai==1.3.7
groq==0.4.1

# Utilities
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2

# Selenium
selenium==4.15.2
webdriver-manager==4.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2